// src/app/models/ticket.ts
export interface Ticket {
    id?:                any;
    dataAbertura?:   string;
    dataFechamento?: string;
    prioridade:      string;
    status:          string;
    titulo:          string;
    observacoes:     string;
    tecnico:            any;
    cliente:            any;
    nomeCliente:     string;
    nomeTecnico:     string;
 }